Isolate the logging when calling the CLI for testing
The CLI configures the logging for the application. Modifying it during
tests adds a lot of garbage to the output and makes tests hard to debug.

This ensures we do not keep around the new handlers, and that the pytest
ones are not in use during that part of the tests.