Reimplement the scheduler to make it more flexible
The previous scheduler was not very nice to work with and required a lot
of duplication in handling state. With this new one, we can now start
making improvements ourselves and remove the various places handling our
state.

This also removes a hard dependency on a python3.9 standard library
dependency, opening the door to supporting python3.8