ci: Limit runtime to 15 minutes and run only one pypy test at a time
Otherwise, when tests hang, they take 6 hours to finish. Our mean time
is 8 minutes, thus 15 should be plenty of headroom

The pypy tests seem to hang on GitHub Actions and we dont' get logs for
what happened. Running them serially seems to gix the problem, let's do
that for now